1-Langchain/1.2-ollama/app.py
import os
from dotenv import load_dotenv
load_dotenv()

from langchain_community.llms import Ollama
import streamlit as st 
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser

##Langsmith Tracing
os.environ["LANGCHAIN_API_KEY"] = os.getenv("LANGCHAIN_API_KEY")
os.environ["LANGCHAIN_TRACING_V2"] = "true" 
os.environ["LANGCHAIN_PROJECT"] = os.getenv("LANGCHAIN_PROJECT")


##Prompt Template 
prompt=ChatPromptTemplate.from_messages(
    [
        ("system","You are a helpful assistant .Please respond to the question asked"),
        ("user","Question:{question}")
    ]
)

##Streamlit Framework
st.title("Langchain Demo With LLAMA3")
input_text=st.text_input("What question ypu have in mind?")


##Ollama Llana2 model
llm=Ollama(model="llama3")
output_parser=StrOutputParser()
chain=prompt | llm | output_parser

if input_text:
    st.write(chain.invoke({"question":input_text}))
