requirements.txt
langchain
langchain-classic
langchain-openai
langchain-community
langchain-text-splitters
python-dotenv
ipykernel
beautifulsoup4
faiss-cpu
streamlit